# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Octo Protocol API
  version: 1.0.0
  description: >
    OpenAPI contract for the /v1 Octo Protocol routes.


    Every `amount_stroops` field below is a JSON `integer` and stays that way in this version —
    that is unchanged by issue #215. Storage now also carries an arbitrary-precision
    `amount_base_units NUMERIC(78,0)` column internally (needed so a future, non-Stellar chain
    whose amounts don't fit 64 bits can be represented exactly), but nothing about that is exposed
    on these schemas yet. The public cutover to a string-typed amount field is a separate,
    versioned breaking change tracked in issue #227 — do not assume `amount_stroops` stays numeric
    across a future major version.
servers:
  - url: /
paths:
  /v1/wallets:
    post:
      summary: Create a master wallet (non-custodial)
      description: >
        Registers a wallet from a keypair the CLIENT generated. The server never receives the
        private key or mnemonic — only the public account. `encrypted_backup` is an opaque blob
        the client encrypted under the user's password; the server stores it verbatim for
        new-device recovery and cannot decrypt it.
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - public_key
              properties:
                public_key:
                  type: string
                  description: The client-generated Stellar account (`G...`).
                encrypted_backup:
                  type: string
                  nullable: true
                  description: >
                    Opaque client-encrypted seed backup. Never decryptable by the server.
                label:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWalletResponse'
  /v1/wallets/{id}:
    get:
      summary: Get wallet details
      operationId: getWallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletViewResponse'
...
  /v1/wallets/{id}/transactions:
    get:
      summary: List transactions
      operationId: listTransactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransactionsResponse'
...
components:
  schemas:
...
    CreateWalletResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - id
            - network
            - address
            - custody
            - funded
          properties:
            id:
              type: string
              format: uuid
            network:
              type: string
            address:
              type: string
            custody:
              type: string
              enum: [client, server]
              description: >
                `client` for non-custodial wallets (the server holds no key for this account).
                `server` is legacy, pre-cutover.
            funded:
              type: boolean
              description: Whether the account was funded on-chain (testnet friendbot).
          # NOTE: `recovery_mnemonic` was previously listed here as REQUIRED. It no longer exists
          # and must not come back: the client generates the mnemonic and the server never sees
          # it. Documenting it would contradict the non-custodial guarantee.
    WalletViewResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - id
            - network
            - address
          properties:
            id:
              type: string
              format: uuid
            network:
              type: string
            address:
              type: string
            label:
              type: string
            description:
              type: string
            balance_stroops:
              type: integer
...
    ListTransactionsResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            type: object
...